requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
cucumber-expressions==18.0.1
cucumber-tag-expressions==8.0.0
Django==5.2.8
django-stubs-ext==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
iniconfig==2.3.0
packaging==25.0
parse==1.20.2
parse_type==0.6.6
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-django==4.11.1
python-decouple==3.8
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
